add up to date kernels, and hopefully fix all bugs with spmv

+++ example/updatekernels.sh
@@ -1,8 +1,18 @@
 #/bin/sh
 
+lift=$1
+
+if [ -z "$1" ]
+  then
+    echo "No lift root supplied, failing"
+    exit 1
+fi
+
+echo "lift root: $lift"
+
 rm *.json
 
-d=/home/adam/projects/lift/scripts/generated_programs/spmv/
+d=$lift/scripts/generated_programs/spmv/
 cp -f $d/swrg-slcl-pmdp.json kernel.json
 cp -f $d/awrg-alcl-alcl-edp-split-512.json kernel2.json
 cp -f $d/swrg-slcl-sdp-chunk-128.json kernel3.json
@@ -15,7 +25,7 @@ for alg in "${algorithms[@]}"
 do 
 	echo "Copying algorithm: $alg"
 	rm -rf $alg
-	ad=/home/adam/projects/lift/scripts/generated_programs/$alg/
+	ad=$lift/scripts/generated_programs/$alg/
 	mkdir -p $alg
 	cp -f $ad/swrg-slcl-pmdp.json $alg/kernel.json
 	cp -f $ad/awrg-alcl-alcl-edp-split-512.json $alg/kernel2.json
